llm_function_call/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain_core
langchain_openai